lot of small changes for .html
the page contents done so far are not changed. I just do some formating for the html file. Instead of slim jquery, I change it to the uncompressed one, in order to use Ajax functions. Addition, I found that the navbar content is not quite reusable, maybe we can consider some new contents, which can be applied in every page in MapOut