testing the user flow and modified some details
Reduce the redundant in main, function_bar html file.
Combined the "edit account info" to the user setting.
The login function is not yet built, but clicking login will goto the user page, which is for testing the flow.
Temporarily remove the "Project" in the navbar, add_project_button is now associated with create_project.html